branch prediction fixed in sideways fenwick

+++ lib/completetreeFLOAT.hpp
@@ -11,7 +11,7 @@
 namespace dense {
 namespace stochastic {
 
-  typedef float Real;
+  typedef double Real;
   /* This complete tree class is meant to be subclassed to build various 
    * data structures.  Because most of those data structures require 
    * updates when values are modified, only const access to elements is
